pyproject.toml
[project]
name = "mlb-predict"
version = "0.1.0"
description = "MLB game win-probability and run-total prediction with calibrated ML."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "pydantic>=2.6",
    "pandas>=2.2",
    "numpy>=1.26",
    "pyarrow>=15.0",
    "python-multipart>=0.0.9",
    "scikit-learn>=1.4",
    "lightgbm>=4.3",
    "shap>=0.45",
    "matplotlib>=3.8",
    "pybaseball>=2.2",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.9"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
